BUGGY: rust porting for mesh and better point propagation system in works

+++ src-tauri/src/notify_block.rs
@@ -1,7 +1,7 @@
 use serde::Serialize;
 use tauri::window::Window;
 
-pub fn notify_block(index: i32, name: &str, window: &Window) -> Result<(), tauri::Error>
+pub fn notify_block(index: usize, name: &str, window: &Window) -> Result<(), tauri::Error>
 {
     let blck_info = BlckInfo{index, name};
     window.emit(name, blck_info)
@@ -9,6 +9,6 @@ pub fn notify_block(index: i32, name: &str, window: &Window) -> Result<(), tauri
 
 #[derive(Serialize, Clone)]
 struct BlckInfo<'a> {
-    index: i32,
+    index: usize,
     name: &'a str,
 }
